Refactor to hexagonal architecture
This commit refactors the project to work with the hexagonal architecture (with setup to use DDD) instead of the MVC classic setup of SpringBoot projects.
