+++ gradle/libs.versions.toml
@@ -2,7 +2,7 @@
 
 coroutines = "1.8.1"
 jooqDockerPlugin = "0.3.12"
-jvmTarget = "17"
+jvmTarget = "21"
 kotlin = "1.9.24"
 mockk = "1.13.10"
 mockkSpring = "4.0.2"
@@ -30,7 +30,7 @@ springboot_jdbc_starter = { module = "org.springframework.boot:spring-boot-start
 springboot_jooq_starter = { module = "org.springframework.boot:spring-boot-starter-jooq" }
 springboot_test_starter = { module = "org.springframework.boot:spring-boot-starter-test" }
 springboot_validation_starter = { module = "org.springframework.boot:spring-boot-starter-validation" }
-springboot_web_starter = { module = "org.springframework.boot:spring-boot-starter-web" }
+springboot_web_starter = { module = "org.springframework.boot:spring-boot-starter-webflux" }
 springdoc_openapi_starter = { module = "org.springdoc:springdoc-openapi-starter-webmvc-ui", version.ref = "springdoc" }
 
 # Persistence
