+++ spring-boot-app-template/Dockerfile
@@ -1,4 +1,4 @@
-FROM eclipse-temurin:17-jre-alpine
+FROM eclipse-temurin:21-jre-alpine
 
 ENV APP_BASE="/home" \
     APP_NAME="spring-boot-app-template" \
