+++ spring-boot-app-template/src/main/kotlin/com/yonatankarp/springbootapptemplate/application/ports/HelloWorldPort.kt
@@ -1,0 +1,5 @@
+package com.yonatankarp.springbootapptemplate.application.ports
+
+interface HelloWorldPort {
+    suspend fun greet(): String
+}
