+++ spring-boot-app-template/src/main/kotlin/com/yonatankarp/springbootapptemplate/domain/valueobject/HelloWorld.kt
@@ -1,0 +1,5 @@
+package com.yonatankarp.springbootapptemplate.domain.valueobject
+
+class HelloWorld {
+    fun sayHello(): String = "Hello, World!"
+}
