+++ spring-boot-app-template/src/test/kotlin/com/yonatankarp/springbootapptemplate/adapters/AbstractIntegrationTest.kt
@@ -1,4 +1,4 @@
-package com.yonatankarp.springbootapptemplate
+package com.yonatankarp.springbootapptemplate.adapters
 
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
